Found bug (i'm sorry i have failed) and also tweaked the possessing_entity condition to make it better

+++ src/main/java/morgan/lesbois/interfaces/DoubleJumpInterface.java
@@ -2,7 +2,7 @@ package morgan.lesbois.interfaces;
 
 public interface DoubleJumpInterface {
     public int lesbois$getMaxDoubleJumps();
-    public int lesbois$getDoubleJumpHeight();
+    public double lesbois$getDoubleJumpHeight();
     public int lesbois$getDoubleJumps();
     public void lesbois$setDoubleJumps(int doubleJumps);
